// types/property.ts
export interface Property {
    id: number;
    name: string;
    location: string;
    priceRange: string;
    bedrooms: string;
    bathrooms: string;
    area: string;
    image: string;
    description: string;
    type: string;
    status: string;
    features: string[];
    amenities: string[];
    gallery: string[];
    isFeatured: boolean;
}